requirements.txt
discord.py==1.4.1
gspread==3.6.0
matplotlib==3.3.3
oauth2client==4.1.3
python-dotenv==0.10.5
PyYAML==5.3
regex==2020.1.8
verbecc==1.5.9
